// src/app/interfaces/books.interface.ts
export interface BooksInterface {
        
        id?: number;
        title: string;
        rating: number;
        authors: string[];
        price: number;
        quantity: number;
        stock:string;
        description: string;
        imageUrl:string;
        showFullDescription: boolean;
    
}